# pkg/handler/docs/swagger.yaml
basePath: /
definitions:
...
  handler.ApiResponse-service_TokenResponse:
    properties:
      data:
        $ref: '#/definitions/service.TokenResponse'
      error:
        type: string
    type: object
  handler.ApiResponse-string:
    properties:
      data:
        type: string
      error:
        type: string
    type: object
...
  service.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
info:
  contact: {}
  description: Authentication service/library for Go.
  title: EzAuth API
  version: "1.0"
paths:
...
  /auth/oauth2/{provider}/callback:
    get:
      description: Handle the callback from the OAuth2 provider
      parameters:
      - description: OAuth2 Provider (google, github, facebook, discord, gitlab, slack,
          linkedin, spotify, or any registered custom provider)
        in: path
        name: provider
        required: true
        type: string
      - description: Authorization Code
        in: query
        name: code
        required: true
        type: string
      - description: CSRF State
        in: query
        name: state
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ApiResponse-service_TokenResponse'
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ApiResponse-string'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ApiResponse-string'
      summary: OAuth2 Callback
      tags:
      - oauth2
  /auth/oauth2/{provider}/login:
    get:
      description: Redirect to the OAuth2 provider login page
      parameters:
      - description: OAuth2 Provider (google, github, facebook, discord, gitlab, slack,
          linkedin, spotify, or any registered custom provider)
        in: path
        name: provider
        required: true
        type: string
      responses:
        "307":
          description: Temporary Redirect
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ApiResponse-string'
      security:
      - ApiKeyAuth: []
      summary: OAuth2 Login
      tags:
      - oauth2
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"